Refactor : QuizRequestDTO의 selectedOptionId type 변경

+++ src/main/java/com/example/cp_main_be/domain/mission/quiz/dto/QuizRequestDTO.java
@@ -9,5 +9,5 @@ import lombok.Setter;
 public class QuizRequestDTO {
 
   private QuizType quizType;
-  private Integer selectedOptionId;
+  private Long selectedOptionId;
 }
